Login Integrated Test
Unit Test added to check whether the username and password exists in the database or not. If it does exist the test checks that screen changes from the login window to the next page. This is the integrated form of the test as it is combining frontend and backend testing for the login feature.It tests both the database as well as the input on the frontend side. The test passes.