Unit Test added to check Invalid Case
The Unit Test checks when the user enters an invalid username and password combination, the test searches the database to check that the combination is not found. The test then clicks on the login button and checks that the screen does not change and remains on the login page. Integrated Login Test - 2. The test passes.