Backend Login Unit Test Added
Unit test added to check if the username and password combination entered by the user exists in the database or not and accordingly checks if the screen changes to the next window. The key difference between this unit test and previously committed integration test is that instead of manually entering a username the test will search the the database, pick a random username entry and then test it. The purpose if this is to focus testing the backend login part instead of the combination of the feature with frontend.